requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0
httpx~=0.28.0
jwt~=1.3.1
python-dotenv~=1.0.1
